+++ src/modules/services/elasticmq.nix
@@ -23,6 +23,15 @@ in
   };
 
   config = lib.mkIf cfg.enable {
-    processes.elasticmq-server.exec = "exec env JAVA_TOOL_OPTIONS=\"-Dconfig.file=${pkgs.writeText "elasticmq-server.conf" cfg.settings}\" ${cfg.package}/bin/elasticmq-server";
+    processes.elasticmq-server = {
+      exec = "exec env JAVA_TOOL_OPTIONS=\"-Dconfig.file=${pkgs.writeText "elasticmq-server.conf" cfg.settings}\" ${cfg.package}/bin/elasticmq-server";
+      ready = {
+        http.get = {
+          path = "/health";
+          port = 9324;
+        };
+        initial_delay = 2;
+      };
+    };
   };
 }